% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_way.R
\name{perform_matrix_operations}
\alias{perform_matrix_operations}
\title{Perform Matrix Operations}
\usage{
perform_matrix_operations(a, b)
}
\arguments{
\item{a}{A numeric matrix.}

\item{b}{Another numeric matrix.}
}
\value{
A list containing:
\itemize{
\item \code{result_ab_t}: the product of \code{a} and the transpose of \code{b}.
\item \code{result_diff1}: the difference between the vectorized product of \code{a} and \code{b^T} and the product of \code{b \%x\% diag(nrow(a))} and \code{a}.
\item \code{result_diff2}: the difference between the product of \code{b \%x\% diag(nrow(a))} and \code{a} and the tensor product of \code{b} and \code{a}.
}
}
\description{
This function performs specific matrix operations including multiplication of a matrix and
the transpose of another and calculates differences based on tensor multiplication with a diagonal matrix.
}
\examples{
a <- matrix(1:10, ncol = 1)
b <- matrix(1:5, ncol = 1)
perform_matrix_operations(a, b)
}
