% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_student_t.R
\name{scr_tucker3_student_parameter_count}
\alias{scr_tucker3_student_parameter_count}
\title{Parameter Count for Student-t Tucker3 SCR}
\usage{
scr_tucker3_student_parameter_count(
  groups,
  variables,
  occasions,
  centroid_rank,
  variable_rank,
  occasion_rank,
  estimate_degrees_freedom = FALSE
)
}
\arguments{
\item{groups}{Number of mixture components G.}

\item{variables}{Number of variables J.}

\item{occasions}{Number of occasions K.}

\item{centroid_rank}{Centroid-mode rank P.}

\item{variable_rank}{Variable-mode rank Q.}

\item{occasion_rank}{Occasion-mode rank R.}

\item{estimate_degrees_freedom}{Logical; add one parameter when nu is
estimated rather than fixed.}
}
\value{
Integer parameter count.
}
\description{
Extend the Gaussian Tucker3 parameter count for a Student-t model.
}
