% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_way.R
\name{tensor_products}
\alias{tensor_products}
\title{Compute Various Tensor Products of Two Vectors}
\usage{
tensor_products(i1, i2)
}
\arguments{
\item{i1}{A numeric vector or matrix to be used in tensor products.}

\item{i2}{A numeric vector or matrix to be used in tensor products.}
}
\value{
A list containing various tensor products:
\itemize{
\item \code{i1xi1xi1}: Tensor product of i1 with itself three times.
\item \code{i1xi1xi2}: Tensor product of i1 with itself twice and then with i2.
\item \code{i1xi2xi1}: Tensor product of i1 with i2 and then with i1.
\item \code{i1xi2xi2}: Tensor product of i1 with i2 twice.
\item \code{i2xi1xi1}: Tensor product of i2 with i1 twice.
\item \code{i2xi1xi2}: Tensor product of i2 with i1 and then with i2.
\item \code{i2xi2xi1}: Tensor product of i2 with itself and then with i1.
\item \code{i2xi2xi2}: Tensor product of i2 with itself three times.
}
}
\description{
This function computes all possible tensor products of two input vectors i1 and i2,
considering combinations of tensoring them up to three times in all possible orders.
}
\examples{
i1 <- matrix(c(1, 0), ncol = 1)
i2 <- matrix(c(0, 1), ncol = 1)
tensor_products(i1, i2)
}
